pyproject.toml
[project]
name = "technical-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.2",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.0.1",
    "langchain-text-splitters>=1.0.0",
    "pdfplumber>=0.11.7",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
